Clening up behaviour state logic. #15
Initialization of agentInvResource:
Added a check to initialize agentInvResource if it is null or undefined.

Modification in isHungry Method:
Changed the condition to check if the amount of food in the agent's inventory is greater than or equal to this.resourceHunger.

Code Cleanup in setTarget Method:
Removed an unnecessary blank line in the setTarget method.